Version 0.03 Top down map of Level 1
The 3D environment is simplified to set up, the floors and have been set for the first level, with walls coming shortly after, and the environment's platforms/objects will come in the same version. This has been updated to 0.03 as the new concept added is a full floorplan of level 1. This phase of development is slow, but the plan is laid out.

P.S use the below video for ideas on interactive
objects for your game :)
https://www.youtube.com/watch?v=GJqbxH1sESY