Version 0.07 Bullets
Bullets have been implemented thanks to the use of a YouTube Tutorial, and I still need to figure out how to apply the bullets to any other wall or enemy for collision. I have also put walls and (hopefully) appropriate mesh roadmaps for enemies. You can still stand on tables to avoid them entirely, which is an oversight.