V 0.09 Enemies Fix 2: seek and die
The enemies need to be adjusted so that the 'mesh instance' in them creates a collision shape in the 'mesh' settings called a 'convex collision sibling' which is automatically connected to the mesh. :)

This means they can be hit. However, a bug with this is that by doing this, they cannot move (probably a bug with the collision shape in the floor). A temporary fix to this is making the enemies shorter (scaling).