Version 0.2.1 Fullscreen Compatibility
Added an option in the options menu to do fullscreen. The window cannot be resized other than this method.

I have plans to meet up with Bailey to inform me on cool stuff to do with the options menu (funny modes that I have heard great appeal for in my user testing).