Version 0.3 technically final development
Options Fix, shows score and how many enemies killed at the end (might be broken in level 2?), and started work on the next level's framework, which might be done on a later date. Documentation in the files of the game have been done, and analytics will only need to be done in google docs. :)